moved the classes under CSV2RDF to separate files
