+++ src/com/complexible/common/csv/ValueGenerator.java
@@ -1,0 +1,7 @@
+package com.complexible.common.csv;
+
+import org.openrdf.model.Value;
+
+public interface ValueGenerator<V extends Value> {
+    V generate(int rowIndex, String[] row);
+}
